[Silver III] Title: 주유소, Time: 24 ms, Memory: 3584 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -2,8 +2,8 @@
 using namespace std;
 using ll = long long;
 int n;
-int dist[100001];
-int cost[100001];
+ll dist[100001];
+ll cost[100001];
 
 int main() {
     cin.tie(0)->sync_with_stdio(0);
@@ -12,7 +12,6 @@ int main() {
     for (int i = 0; i < n; i++) cin >> cost[i];
     ll ret = cost[0] * dist[0];
     ll minCost = cost[0];
-
     for(int i = 1; i < n-1; i++){
         if(cost[i] < minCost){
             minCost = cost[i];
